move features to world.gen.feature
