+++ src/main/java/net/newfrontiercraft/nfc/mixin/SaplingBlockMixin.java
@@ -3,15 +3,14 @@ package net.newfrontiercraft.nfc.mixin;
 import net.minecraft.block.Block;
 import net.minecraft.block.PlantBlock;
 import net.minecraft.block.SaplingBlock;
-import net.minecraft.block.material.Material;
 import net.minecraft.world.World;
 import net.minecraft.world.biome.Biome;
 import net.minecraft.world.gen.feature.*;
 import net.newfrontiercraft.nfc.events.init.BlockListener;
-import net.newfrontiercraft.nfc.feature.PlanterBirchTreeFeature;
-import net.newfrontiercraft.nfc.feature.PlanterOakTreeFeature;
-import net.newfrontiercraft.nfc.feature.PlanterSpruceTreeFeature;
-import net.newfrontiercraft.nfc.feature.ShrubFeature;
+import net.newfrontiercraft.nfc.world.gen.feature.PlanterBirchTreeFeature;
+import net.newfrontiercraft.nfc.world.gen.feature.PlanterOakTreeFeature;
+import net.newfrontiercraft.nfc.world.gen.feature.PlanterSpruceTreeFeature;
+import net.newfrontiercraft.nfc.world.gen.feature.ShrubFeature;
 import net.newfrontiercraft.nfc.world.gen.feature.DeadTreeFeature;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
