+++ src/main/java/net/newfrontiercraft/nfc/world/gen/feature/ScorchedSandFeature.java
@@ -1,4 +1,4 @@
-package net.newfrontiercraft.nfc.feature;
+package net.newfrontiercraft.nfc.world.gen.feature;
 
 import net.minecraft.block.Block;
 import net.minecraft.util.math.MathHelper;
